xWarnCanDeleteWarning hook, [xAdmin] -> [xWarn], force reason & use targetply instead of name

+++ lua/xwarn/commands/sv_warn.lua
@@ -9,12 +9,17 @@ xAdmin.Core.RegisterCommand("warn", "Warn a user", 30, function(admin, args)
 	local target, targetPly = xAdmin.Core.GetID64(args[1], admin)
 
 	if not target then
-		xAdmin.Core.Msg({xAdmin.Config.ColorLog, "[xAdmin] ", color_white, "Please provide a valid target. The following was not recognised: " .. args[1]}, admin)
+		xAdmin.Core.Msg({xAdmin.Config.ColorLog, "[xWarn] ", color_white, "Please provide a valid target. The following was not recognised: " .. args[1]}, admin)
 
 		return
 	end
 
-	local reason = args[2] or "No reason given"
+	local reason = args[2]
+
+	if not reason then 
+		xAdmin.Core.Msg({xAdmin.Config.ColorLog, "[xWarn] ", color_white, "Please provide a reason."}, admin)
+		return
+	end
 
 	if args[3] then
 		for k, v in pairs(args) do
@@ -24,7 +29,7 @@ xAdmin.Core.RegisterCommand("warn", "Warn a user", 30, function(admin, args)
 	end
 
 	xWarn.Database.CreateWarn(target, (IsValid(targetPly) and targetPly:Name()) or "Unknown", admin:SteamID64(), admin:Name(), reason or "No reason given")
-	xAdmin.Core.Msg({admin, " warned ", ((IsValid(targetPly) and targetPly:Name()) or target), " for: ", Color( 255, 0, 0, 255 ), reason})
+	xAdmin.Core.Msg({admin, " warned ", ((IsValid(targetPly) and targetPly) or target), " for: ", Color( 255, 0, 0, 255 ), reason})
 	hook.Run("xWarnPlayerWarned", target, targetPly, admin, reason)
 end)
 
@@ -39,10 +44,11 @@ xAdmin.Core.RegisterCommand("warns", "View a user's warnings", 30, function(admi
 	local target, targetPly = xAdmin.Core.GetID64(args[1], admin)
 
 	if not target then
-		xAdmin.Core.Msg({xAdmin.Config.ColorLog, "[xAdmin] ", color_white, "Please provide a valid target. The following was not recognised: " .. args[1]}, admin)
+		xAdmin.Core.Msg({xAdmin.Config.ColorLog, "[xWarn] ", color_white, "Please provide a valid target. The following was not recognised: " .. args[1]}, admin)
 
 		return
 	end
+
 	xWarn.Database.GetWarns(target, function(warns)
 		if warns == nil then xWarn.msg("You have no warnings!", user) return end
 		for k, v in pairs(warns) do
@@ -79,16 +85,21 @@ xAdmin.Core.RegisterCommand("deletewarn", "Delete a warning (by ID)", 40, functi
 		return
 	end
 	if (not tonumber(args[1])) then
-		xAdmin.Core.Msg({Color(46, 170, 200), "[xAdmin] ", color_white, "'" .. args[1] .. "' is not an ID. Please provide a valid ID."}, admin)
+		xAdmin.Core.Msg({Color(46, 170, 200), "[xWarn] ", color_white, "'" .. args[1] .. "' is not an ID. Please provide a valid ID."}, admin)
 		return
 	end
 	xWarn.Database.GetWarnById(args[1], function(warn)
 		if warn and warn[1] then
+			local canDelete, msg = hook.Run("xWarnCanDeleteWarning", admin, warn[1])
+			if canDelete == false then
+				xAdmin.Core.Msg({Color(46, 170, 200), "[xWarn] ", color_white, msg}, admin)
+				return
+			end
 			xWarn.Database.DestroyWarn(args[1])
 			xAdmin.Core.Msg({admin, " deleted warning with ID ", args[1]})
 			hook.Run("xWarnWarningDeleted", args[1], admin)
 		else 
-			xAdmin.Core.Msg({Color(46, 170, 200), "[xAdmin] ", color_white, "'" .. args[1] .. "' is not a valid ID. Please provide a valid ID."}, admin)
+			xAdmin.Core.Msg({Color(46, 170, 200), "[xWarn] ", color_white, "'" .. args[1] .. "' is not a valid ID. Please provide a valid ID."}, admin)
 		end
 	end )
 end)
